feat: 비디오 모달 기능 추가 및 Calendar 클래스에서 비디오 처리 로직 통합
